Move DB staff to another module to speed up building
