+++ .gitignore
@@ -12,3 +12,4 @@
 /app/src/main/assets/web/.sourcemaps/
 /app/src/main/assets/web/loader.js
 /app/bibleview-js/dist/
+/db/build/
