+++ app/src/main/java/net/bible/android/control/page/CurrentBiblePage.kt
@@ -20,9 +20,9 @@ package net.bible.android.control.page
 import android.app.Activity
 import android.util.Log
 import net.bible.android.control.versification.BibleTraverser
+import net.bible.android.database.workspaces.WorkspaceEntities
 import net.bible.android.view.activity.navigation.GridChoosePassageBook
 import net.bible.service.common.CommonUtils.getWholeChapter
-import net.bible.service.db.workspaces.WorkspaceEntities
 import net.bible.service.sword.SwordContentFacade
 import net.bible.service.sword.SwordDocumentFacade
 import org.apache.commons.lang3.StringUtils
