+++ app/src/main/java/net/bible/android/control/page/CurrentPage.kt
@@ -20,7 +20,7 @@ package net.bible.android.control.page
 import android.app.Activity
 import android.view.Menu
 import net.bible.service.common.ParseException
-import net.bible.service.db.workspaces.WorkspaceEntities
+import net.bible.android.database.workspaces.WorkspaceEntities
 import net.bible.service.format.Note
 import org.crosswire.jsword.book.Book
 import org.crosswire.jsword.book.BookCategory
