+++ app/src/main/java/net/bible/android/control/page/CurrentPageBase.kt
@@ -22,7 +22,7 @@ import android.view.Menu
 import net.bible.android.activity.R
 import net.bible.android.control.PassageChangeMediator
 import net.bible.service.common.ParseException
-import net.bible.service.db.workspaces.WorkspaceEntities
+import net.bible.android.database.workspaces.WorkspaceEntities
 import net.bible.service.format.HtmlMessageFormatter.Companion.format
 import net.bible.service.format.Note
 import net.bible.service.sword.SwordContentFacade
