+++ db/build.gradle
@@ -1,0 +1,74 @@
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
+
+apply plugin: 'kotlin-kapt'
+// apply plugin: 'kotlinx-serialization'
+
+
+android {
+    compileSdkVersion 29
+    buildToolsVersion "29.0.2"
+
+    /** these config values override those in AndroidManifest.xml.  Can also set versionCode and versionName */
+    defaultConfig {
+        minSdkVersion 19
+        targetSdkVersion 29
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled true
+            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+        }
+    }
+}
+
+
+dependencies {
+//    implementation 'androidx.appcompat:appcompat:1.1.0'
+//    implementation 'androidx.drawerlayout:drawerlayout:1.0.0'
+//    implementation 'androidx.media:media:1.1.0'
+//    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+//    implementation 'androidx.core:core-ktx:1.1.0'
+
+//    implementation 'com.google.android.material:material:1.0.0'
+//    implementation "org.jetbrains.anko:anko-commons:0.10.8"
+
+    // To support Android 4.4
+//    implementation 'androidx.multidex:multidex:2.0.1'
+    
+
+    // allow annotations like UIThread, StringRes see: https://developer.android.com/reference/android/support/annotation/package-summary.html
+//    implementation 'androidx.annotation:annotation:1.1.0'
+
+    implementation "androidx.room:room-runtime:$room_version"
+    kapt "androidx.room:room-compiler:$room_version" // For Kotlin use kapt instead of annotationProcessor
+    //annotationProcessor "androidx.room:room-compiler:$room_version" // For Kotlin use kapt instead of annotationProcessor
+
+    // optional - Kotlin Extensions and Coroutines support for Room
+    implementation "androidx.room:room-ktx:$room_version"
+
+    // optional - RxJava support for Room
+    // implementation "androidx.room:room-rxjava2:$room_version"
+
+    // optional - Guava support for Room, including Optional and ListenableFuture
+    // implementation "androidx.room:room-guava:$room_version"
+
+    // Test helpers
+    // testImplementation "androidx.room:room-testing:$room_version"
+
+
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    implementation "org.jetbrains.kotlinx:kotlinx-serialization-runtime:0.11.0"
+
+//    implementation 'com.google.dagger:dagger:2.25.2'
+//    annotationProcessor 'com.google.dagger:dagger-compiler:2.25.2'
+//    kapt 'com.google.dagger:dagger-compiler:2.25.2'
+//
+//    implementation 'de.greenrobot:eventbus:2.4.1'
+}
+
+repositories {
+	mavenCentral()
+}
